Moving files to setup, adding rhai scripts to plugin
