+++ plugins/shrs_completion/examples/completer.rs
@@ -2,8 +2,8 @@ use shrs::prelude::*;
 use shrs_completion::completions::*;
 
 fn main() {
-    let mut mycompleter = Completer::default();
-    ssh_completion(&mut mycompleter);
+    let mut mycompleter = DefaultCompleter::default();
+    mycompleter.register(ssh_rule());
 
     let myshell = ShellBuilder::default().build().unwrap();
