+++ plugins/shrs_completion/src/completions/ssh.rs
@@ -7,8 +7,8 @@ use crate::helpers::known_hosts;
 
 // TODO this only works with the default completer, perhaps make register part of the completer
 // trait?
-pub fn ssh_completion(comp: &mut Box<dyn Completer>) {
-    comp.register(Rule::new(Pred::new(ssh_pred), Box::new(known_hosts_action)));
+pub fn ssh_rule() -> Rule {
+    Rule::new(Pred::new(ssh_pred), Box::new(known_hosts_action))
 }
 
 fn ssh_pred(ctx: &CompletionCtx) -> bool {
